+++ app/src/calibration/constants.js
@@ -5,3 +5,6 @@ export const START_DECK_CHECK: 'calibration:START_DECK_CHECK' =
 
 export const START_DECK_CHECK_SUCCESS: 'calibration:START_DECK_CHECK_SUCCESS' =
   'calibration:START_DECK_CHECK_SUCCESS'
+
+export const DECK_CHECK_PATH: '/calibration/check/session' =
+  '/calibration/check/session'
